+++ requests.go
@@ -697,11 +697,39 @@ type QuerySegmentsRequest struct {
 	// Array of sort parameters
 	Sort []SortParamRequest `json:"sort"`
 }
+type CreateSegmentRequest struct {
+	// The type of the segment
+	Type string `json:"type"`
+	// If true, all sender channels are included in the segment
+	AllSenderChannels *bool `json:"all_sender_channels,omitempty"`
+	// If true, all users are included in the segment
+	AllUsers *bool `json:"all_users,omitempty"`
+	// The description of the segment (max 256 characters)
+	Description *string `json:"description,omitempty"`
+	// The ID of the segment
+	ID *string `json:"id,omitempty"`
+	// The name of the segment (max 128 characters)
+	Name *string `json:"name,omitempty"`
+	// Filter to apply to the query
+	Filter map[string]any `json:"filter"`
+}
+type UpdateSegmentRequest struct {
+	// The description of the segment (max 256 characters)
+	Description *string `json:"description,omitempty"`
+	// The name of the segment (max 128 characters)
+	Name *string `json:"name,omitempty"`
+	// Filter to apply to the query
+	Filter map[string]any `json:"filter"`
+}
 type DeleteSegmentRequest struct {
 }
 type GetSegmentRequest struct {
 }
 type DeleteSegmentTargetsRequest struct {
+	// Target IDs
+	TargetIds []string `json:"target_ids"`
+}
+type AddSegmentTargetsRequest struct {
 	// Target IDs
 	TargetIds []string `json:"target_ids"`
 }